Reduce Saleable's footprint using presenter objects
Saleable does a number of things, including being the "glue" between the
Spree::Product and the catalog record, but it was also used to transform
attributes between the two records on creation. We're now doing that
with the WaxPoetic::Product objects, stored in `./app/products`, which
provide a cleaner and less complicated interface for presenting data for
creation as a Spree::Product. All of the defined methods are eventually boiled
down to the `attributes()`, where they are mass-assigned into a
Spree::Product and saved.